Move IDE logic into its own hook
