+++ frontend/src/components/ide/ProgramIde.tsx
@@ -1,23 +1,22 @@
-import React, { useState, useEffect, useCallback, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import { RelayProp, createFragmentContainer } from 'react-relay';
 import graphql from 'babel-plugin-relay/macro';
 import { ProgramIde_hardwareSpec } from './__generated__/ProgramIde_hardwareSpec.graphql';
 import { makeStyles } from '@material-ui/core';
 import CodeEditor from './CodeEditor';
 import RegistersInfo from './RegistersInfo';
-import { IdeContextType, IdeContext, CompiledState } from 'state/ide';
+import { IdeContextType, IdeContext } from 'state/ide';
 import IoInfo from './IoInfo';
 import StackInfo from './StackInfo';
 import IdeControls from './IdeControls';
 import ProgramStatus from './ProgramStatus';
-import { CompileResult, CompilerWrapper } from 'util/compile';
-import { HardwareSpec, ProgramSpec } from 'gdlk_wasm';
 import useDebouncedValue from 'hooks/useDebouncedValue';
 import { assertIsDefined } from 'util/guards';
 import NotFoundPage from 'components/NotFoundPage';
 import { StorageHandler } from 'util/storage';
 import useStaticValue from 'hooks/useStaticValue';
 import PromptOnExit from 'components/common/PromptOnExit';
+import useCompiler from './useCompiler';
 
 const useLocalStyles = makeStyles(({ palette, spacing }) => {
   const border = `2px solid ${palette.divider}`;
@@ -83,105 +82,6 @@ const ProgramIde: React.FC<{
   const userProgram = programSpec.userProgram;
   assertIsDefined(userProgram);
 
-  // These values come from wasm. They're read only, so they're safe to share
-  // with the whole component tree. They are pointers and therefore updates
-  // won't trigger re-renders, but these values shouldn't be changing while
-  // this component tree is mounted anyway.
-  const wasmHardwareSpec = useStaticValue(
-    () =>
-      new HardwareSpec(
-        hardwareSpec.numRegisters,
-        hardwareSpec.numStacks,
-        hardwareSpec.maxStackLength
-      )
-  );
-  const wasmProgramSpec = useStaticValue(
-    () =>
-      new ProgramSpec(
-        Int32Array.from(programSpec.input),
-        Int32Array.from(programSpec.expectedOutput)
-      )
-  );
-
-  // This wasm value is NOT safe to share outside this component. It's stored
-  // in a ref because it contains pointers, which often don't reflect changed
-  // values to React. As such, making it a state field would be useless. This
-  // value has to be manually transformed into compiledState after changes.
-  const compileResult = useRef<CompileResult | undefined>();
-
-  // This is the safe version of the wasm values, which CAN be shared
-  const [compiledState, setCompiledState] = useState<
-    CompiledState | undefined
-  >();
-
-  /**
-   * A manual function called to refresh the shared state that's derived from
-   * the wasm compiled state. This needs to be called explicitly because there
-   * are values in wasm hidden behind pointers. When the values change, the
-   * pointers stay the same, so React doesn't recognize that a change has
-   * occurred. By copying this state into our own JS objects, we allow React to
-   * do its normal updates whenever there's changes.
-   *
-   * @param newCompileResult The new value of compilation output. If not given
-   */
-  const updateCompiledState = useCallback(
-    (newCompileResult: CompileResult | undefined): void => {
-      // Update refs
-      compileResult.current = newCompileResult;
-
-      switch (newCompileResult?.type) {
-        case 'compiled':
-          setCompiledState({
-            type: 'compiled',
-            instructions: newCompileResult.instructions,
-            machineState: newCompileResult.machine.state,
-          });
-          break;
-        case 'error':
-          setCompiledState({
-            type: 'error',
-            errors: newCompileResult.errors,
-          });
-          break;
-        case undefined:
-          setCompiledState(undefined);
-          break;
-      }
-    },
-    [compileResult]
-  );
-
-  /**
-   * Compile the given source code, and update the compiled state with the
-   * new compiled program.
-   *
-   * @param source The source code to compile
-   */
-  const compile = useCallback(
-    (source: string): void => {
-      updateCompiledState(
-        CompilerWrapper.compile(wasmHardwareSpec, wasmProgramSpec, source)
-      );
-    },
-    [wasmHardwareSpec, wasmProgramSpec, updateCompiledState]
-  );
-
-  const execute = useCallback(
-    (executeAll: boolean = false): void => {
-      if (compileResult.current?.type === 'compiled') {
-        compileResult.current.machine.execute(executeAll);
-        // We need to manually refresh since the wasm pointers won't change
-        updateCompiledState(compileResult.current);
-      } else {
-        // This indicates an FE bug, where we tried to step when not allowed
-        throw new Error(
-          'Program is not compiled, cannot execute next instruction.'
-        );
-      }
-    },
-    [compileResult, updateCompiledState]
-  );
-
   // This will be used to read and write source from/to browser local storage
   const sourceStorageHandler = useStaticValue<StorageHandler<string>>(
     () =>
@@ -202,6 +102,14 @@ const ProgramIde: React.FC<{
     return userProgram.sourceCode;
   });
 
+  const {
+    wasmHardwareSpec,
+    wasmProgramSpec,
+    compiledState,
+    compile,
+    execute,
+  } = useCompiler({ hardwareSpec, sourceCode });
+
   // When the source changes, save it to local storage and recompile
   // Use a debounce to prevent constant recompilation
   const debouncedSourceCode = useDebouncedValue(sourceCode, 250);
@@ -209,13 +117,6 @@ const ProgramIde: React.FC<{
     sourceStorageHandler.set(debouncedSourceCode);
     compile(debouncedSourceCode);
   }, [sourceStorageHandler, debouncedSourceCode, compile]);
-
-  // When either spec or the source changes, invalidate the compiled program
-  useEffect(() => {
-    // Do this as a post-effect so that it doesn't run on first render. That
-    // prevents us wiping out state right after we compile
-    return () => updateCompiledState(undefined);
-  }, [wasmHardwareSpec, wasmProgramSpec, sourceCode, updateCompiledState]);
 
   const contextValue: IdeContextType = {
     wasmHardwareSpec,
